Create Workers module and move Event worker there
In preparation of creating 2 more workers which would clutter the main
namespace.

The EventWorker is kept for backwards compatibility. The class can
be later removed after the gems have been upgraded everywhere.
