+++ lib/materialist/event_worker.rb
@@ -1,41 +1,11 @@
-require 'sidekiq'
-require 'active_support/inflector'
+require_relative 'workers/event'
 
+# This class is here for backwards compatibility with pre 3.1 versions. It can be removed with the
+# next major version (4.0)
 module Materialist
-  class EventWorker
-    include Sidekiq::Worker
-
+  class EventWorker < Workers::Event
     def perform(event)
-      topic = event['topic']
-      action = event['type'].to_sym
-      timestamp = event['t']
-
-      materializer = "#{topic.to_s.singularize.classify}Materializer".constantize
-      materializer.perform(event['url'], action)
-
-      report_latency(topic, timestamp) if timestamp
-      report_stats(topic, action, :success)
-    rescue
-      report_stats(topic, action, :failure)
-      raise
-    end
-
-    private
-
-    def report_latency(topic, timestamp)
-      t = (Time.now.to_f - (timestamp.to_i / 1e3)).round(1)
-      Materialist.configuration.metrics_client.histogram(
-        "materialist.event_latency",
-        t,
-        tags: ["topic:#{topic}"]
-      )
-    end
-
-    def report_stats(topic, action, kind)
-      Materialist.configuration.metrics_client.increment(
-        "materialist.event_worker.#{kind}",
-        tags: ["action:#{action}", "topic:#{topic}"]
-      )
+      super(event)
     end
   end
 end
